+++ packages/core/docs/changelog/5939.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Make `register` params in `useUser` optional',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/5939',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Filip Sobol',
-  linkToGitHubAccount: 'https://github.com/filipsobol'
-};