+++ packages/core/docs/commercetools/changelog/5953.js
@@ -1,8 +1,0 @@
-module.exports = {
-  description: 'Fix hardcoded `USD` currency in GraphQL query',
-  link: 'https://github.com/vuestorefront/vue-storefront/pull/5953',
-  isBreaking: false,
-  breakingChanges: [],
-  author: 'Filip Sobol',
-  linkToGitHubAccount: 'https://github.com/filipsobol'
-};